rename SignInView to InitialStudioView
