+++ src/views/studio/InitialStudioView/index.ts
@@ -1,0 +1,3 @@
+import InitialStudioView from './InitialStudioView'
+
+export default InitialStudioView
