+++ src/views/studio/SignInProccessView/index.ts
@@ -1,0 +1,3 @@
+import SignInProccessView from './SignInProccessView'
+
+export default SignInProccessView
