--- src/views/studio/SignInView/index.ts
@@ -1,3 +1,0 @@
-import SignInView from './SignInView'
-
-export default SignInView
